services/mitigation/requirements.txt
# FairOps Mitigation — Dependencies
# Pinned versions from AGENT.md Section 20

fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
httpx==0.27.0

# Fairness / Mitigation
fairlearn==0.10.0
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0

# GCP
google-cloud-aiplatform==1.43.0
google-cloud-bigquery==3.17.0
google-cloud-logging==3.9.0
google-cloud-secret-manager==2.18.0

opentelemetry-sdk==1.22.0
